Follow Purs filenames when generating files
Changes the filename scheme from this:
- `output/src/Module_Name/Module_Name.cpp/.h`
To this:
- `output/Module.Name/index.cpp/.h`

IE: no longer uses a /src subdirectory, and puts files in the existing
`Module.Name` directories created by the purs compiler, instead of
inventing a new naming scheme. The files are named index.cpp/.h: this
corresponds to the index.js javascript files normally created by the
purs compiler. A file named `index` doesn't have the same significance
in c++ as in javascript, but this name was maintained in the absence of
a better one.